feat(user): 프로필 수정 API 구현

+++ src/main/java/com/mini8/backend/features/user/domain/dto/ProfileRequestDTO.java
@@ -18,7 +18,7 @@ public class ProfileRequestDTO {
 
   private Integer careerYears;
 
-  private List<Integer> haveSkillIds;
+  private List<Long> haveSkillIds;
 
-  private List<Integer> wantSkillIds;
+  private List<Long> wantSkillIds;
 }
